+++ src/Index.js
@@ -8,7 +8,7 @@ module.exports = {
   DatePicker: require('./components/DatePicker'),
   DatePickerFullScreen: require('./components/DatePickerFullScreen'),
   DateRangePicker: require('./components/DateRangePicker'),
-  DateRangePicker2: require('./components/DateRangePicker2'),
+  DateRangePickerTest: require('./components/DateRangePickerTest'),
   DateTimePicker: require('./components/DateTimePicker'),
   DisplayInput: require('./components/DisplayInput'),
   DonutChart: require('./components/DonutChart'),
